SY-4786: Make the downsample factor an unsigned 32-bit integer
uint32 matches the lineplot Downsample field that feeds it and maps onto a
proto uint32. The type replaces the streamer's non-negative validation, so a
negative factor is now rejected by the codec rather than by a validator.

+++ core/pkg/distribution/framer/pb/translator.go
@@ -183,7 +183,7 @@ func (IteratorRequestTranslator) Backward(
 		Keys:             channel.KeysFromUint32(req.Keys),
 		ChunkSize:        req.ChunkSize,
 		SeqNum:           int(req.SeqNum),
-		DownsampleFactor: int(req.DownsampleFactor),
+		DownsampleFactor: req.DownsampleFactor,
 	}, nil
 }
 
@@ -204,7 +204,7 @@ func (IteratorRequestTranslator) Forward(
 		Keys:             req.Keys.Uint32(),
 		ChunkSize:        req.ChunkSize,
 		SeqNum:           int32(req.SeqNum),
-		DownsampleFactor: int32(req.DownsampleFactor),
+		DownsampleFactor: req.DownsampleFactor,
 	}, nil
 }
